cliente vendedor delete

+++ src/Controller/ClienteController.php
@@ -121,6 +121,13 @@ class ClienteController extends AbstractController
      */
     public function delete(Request $request, Cliente $cliente): Response
     {
+        $user = $this->security->getUser()->getId();
+
+        if ($cliente->getUser()->getId() !== $user) {
+            return new JsonResponse("404", JsonResponse::HTTP_NOT_FOUND);
+        }
+
+
         if ($this->isCsrfTokenValid('delete'.$cliente->getId(), $request->request->get('_token'))) {
             $em = $this->getDoctrine()->getManager();
             $em->remove($cliente);
